fix(lsp): only use string `Diagnostic.message`s (semgrep/semgrep-proprietary#3186)
Diagnostic messages can only be MarkupContent in LSP 3.18 and later.
While neovim supports this today, VSCode notably does not, and there are
some bundler issues which need to be resolved to use the pre-release of
the 3.18 client implementation in the VSCode extension. Once we're able
to easily support this for VSCode, we should update this to use the
appropriate capability
(`DiagnosticClientCapabilities.markupMessageSupport`) since it appears
`Lsp` doesn't do so automatically.

synced from Pro 23f8e0bfc116cac92e4cf943a6b797f73b70bbeb