Pass test_target_selection with --use-semgrepignore-v2 (semgrep/semgrep-proprietary#3254)
The end-to-end tests in test_target_selection.py now run with and
without `--semgrepignore-v2`.

The default behavior is still to use the legacy implementation which is
equivalent to passing `--no-semgrepignore-v2`.

OCaml support for listing only the files that changed since a base
commit is missing, so we fall back to the Python implementation in such
cases. I'll see if I can get this done soon in another PR. It looks like
a good amount of work. This will allow `--semgrepignore-v2` to work as
promised.

synced from Pro 4bbb61002e66e071a9adc905fded372743e6f068