chore: stop sending errors to CI (semgrep/semgrep-proprietary#3291)
There are some repos wherein the error field of the results is so large that the complete scan endpoint fails, so this will fix that. We now report errors in the text output, and via opentelemetry, so we will still have access to them.

## Test plan
mypy passes. Vibe checked the code.

synced from Pro 6f05dd9690dfda7d15a3df4574c4089001367328