chore: simplify OSS steps (semgrep/semgrep-proprietary#3274)
Previously we were calling a slightly different set of tests in OSS than in pro. In OSS we time `make test` and also run tests in semgrep-interfaces. Now we do both of those in pro too.

Additionally this PR adds a script that makes it easy to change the "root" of the GHA. So when we want to run OSS only things on GHA, we can use this script to do so, and then we won't have to CD into every directory. This should make it easier to port OSS workflows also.

synced from Pro ce27fc46ee37380b3444631cb0564e97285ace92