fix(windows): fix Filename_.readable for windows (semgrep/semgrep-proprietary#3265)
This (deprecated) function used naive regexs to reason about paths
previously, and so was not portable to windows. This change replaces
this logic with equivalent analysis via portable reasoning via `Fpath`.

That there is a note on this function advising use of the `Ppath` in its
place, but that would be a significantly more invasive change. It should
be considered in a followup.

This is a refactor anticipating Windows support.

Test plan:

- Local tests on Windows show that these changes avoid hitting the
`failwith` cases.
- All current tests on other platforms should continue to behave as
before.
- This will help bring us closer to running the test suite on Windows.

synced from Pro 1230e90603e547e2e971e4b7015dd54f273183fe