feat: allow running OSS start-release from pro (semgrep/semgrep-proprietary#3382)
closes SAF-1864, SAF-1865, and SAF-1866

This PR factors the oss release workflow such that it can be run from pro. Luckily most of the steps I could easily plug in the repo name and same w.r.t. for the checks that wait on jobs to finish. 

I also added an input that is the sha hash of what commit specifically to release. This will allow us in a future PR to run this workflow from the pro release workflow, as users will just have to put in two shas (one for pro, one for OSS), when running said workflow. Hopefully this also helps with @nmote's planned work too.

I will update the on call release handbook once this PR is merged to say:
* run pro bump version (and copy hash of merged PR)
* run pro release workflow and paste
* sync pro to oss (and copy hash of merged PR)
* run trigger release semgrep in PRO and paste above hash

The only difference above from the current steps is that the trigger release workflow is now in PRO and requires a hash 

It seems like it'll "just work" (famous last words)

Next steps:
* remove the OSS workflow
* use this workflow from the pro release workflow so users do not need to run two workflows

## Test plan
once the previous PR is merged I should be able to run the modified workflow from this branch, and will do so as a dry run. I am trusting that the dry run process will flex the necessary code paths for an actual release.

See this [workflow run](https://github.com/semgrep/semgrep-proprietary/actions/runs/13823632499/workflow) for where this test succeeded. I am a bit skeptical this tested enough though...

After this PR is merged I will offer to do the release process for whoever is on call then, so I can deal with any issues.

synced from Pro 01748a81b7d9c4f354cc44b957fe0344f92eb124