Warn of upcoming differences with Semgrepignore v2 (semgrep/semgrep-proprietary#3470)
This collects target files using both the Semgrepignore v1
implementation (Python) and the Semgrepignore v2 implementation
(OCaml/Python). If the user is already using Semgrepignore v2, nothing
is printed. If they're using Semgrepignore v1, which is still the
default for all users, one of two warnings is shown:

* if there are no differences in the selected target files, we still
print a message announcing Semgrepignore v2;
* if there are differences in the selected target files, we print a
warning and we list some of the files that are gone and some of the
files that are new in v2 (up to 10 files, 1000 in verbose mode)

Starting with next week's release (1.115.0), everyone will see these
messages unless they force v2 with `--semgrepignore-v2`. If things go
well, we'll change the default from v1 to v2 after that. The contents of
these messages were pre-approved by @emjin, @milanwilliams, and
@mjambon.

This is a screenshot of a sample semgrep run on my machine
(semgrep-proprietary repo, work folder = `OSS/`):


![image](https://github.com/user-attachments/assets/1e27879a-a674-4e8a-b82c-9b594558909d)

Test plan: the existing test snapshots already include the three kinds
of output (v1 no diffs, v1 some diffs, v2) so I didn't add new tests.
For manual testing, run `semgrep scan` or `semgrep ci` on the project of
your choice.

Implementation: the number of selected targets that is reported is the
number of targets that were analyzed by at least one rule. These target
paths were accumulated using a mutable set of type `Set[Path]`. I
changed this type to `TargetAccumulator` which is an object that holds
both the v1 targets and the v2 targets. The actual targets that are used
depend on the `use_semgrepignore_v2` setting and are available as
`.targets()`.

synced from Pro 2aba165e24c35ef353166fa982e25ef20284ed75