Implement timeouts on Windows using GC Memprof (statmemprof) (semgrep/semgrep-proprietary#3201)
The semgrep process has timeouts to cancel a possibly long-running
computation. It implements the timeout by using
[`Unix.setitimer`](https://pubs.opengroup.org/onlinepubs/9699919799/functions/setitimer.html),
which asks the OS to send a signal after a timeout. The signal is then
processed in a signal handler, which raises, and cancels the
computation. What's nice is that timers cannot be nested, and there's at
most one timer running at the same time.

OCaml's libunix doesn't let us send signals on Windows, so we need a
workaround. We could send the long-running computation to a secondary
thread or domain, but unfortunately there's no way of implementing
cancellation with domains or threads, and no way of preempting the
execution to check how much time has elapsed. The process can't just be
exit'ed if the timeout expires as it's supposed to execute the next task
even if a timeout occurs.

The remaining possibility is to use a GC alarm. We can reasonably expect
that a long-running computation will eventually trigger a major
collection. In the GC alarm, we can check how much time has elapsed
since the start of the computation, and raise an `exception Timeout` if
needed. There's a fear that GC alarms won't be triggered frequently
enough, so we use
[`GC.Memprof`](https://ocaml.org/manual/5.3/api/Gc.Memprof.html)
(restored in OCaml 5.3 with backports to 5.2) with a low sampling rate,
which should have no visible effect on performance, on Windows.

The last option would be to convert the semgrep process to use a
cooperative scheduler such as Eio, to use its cancellation mechanism.

I've tested this code with `timeout.js`, and each task timeout around ~2
seconds as expected. Running with a really short timeout also exits
fast, as expected.
`./bin/semgrep.exe --timeout=0.1 timeout.js`

Note that `setitimer` and `getitimer` have been removed in POSIX Issue 8
(2024). See instead
[`timer_settime`](https://pubs.opengroup.org/onlinepubs/9799919799/functions/timer_settime.html).

Test plan:

- All current tests should continue to pass in CI
- You can run `semgrep --pro ci` on Windows without having to add
workarounds to disable the timeout.
- You can run with timeouts on Windows and observe that they take
effect.

We have followup work in #3343 and related PRs to verify the correctness
in CI on Windows.

synced from Pro 559fc4984bf65aab37cff9fb09b49d397e1b7a30