feat: build/test OSS windows wheel (semgrep/semgrep-proprietary#3568)
closes SAF-1929

Previously we were building a windows x86 wheel in the pro repo for testing, but it include the pro binary. This is probably fine to use, but let's refactor a bit and build a normal wheel for uploading, and keep the pro one around for testing.

## Test plan
CI passes

synced from Pro 72674a1611e468098352c0392ca423719c15ad1d