chore: only build/test oss docker image in OSS (semgrep/semgrep-proprietary#3689)
Previously we would target `semgrep-cli`. This would install semgrep-pro. The downside here is this means we cannot do a sync in between the version bump and pro release, as the pro release updates a manifest which is needed by semgrep to download the corresponding pro version (which does not exist yet as pro release has not run).

So right now the release steps are:
1. version bump
2. pro release
3. sync to oss
4. oss release

where 3 must come after 2.

Since we are now publishing the semgrep/semgrep docker image from pro, it's no big deal if we don't test this (But let's leave it for others just in case). This will allow us to do the following:

1. version bump
a. automatically sync to oss
2. pro release
a. automatically trigger oss release

so things will be much simpler.

## Test plan
CI passes

synced from Pro 7e9c0c319fc7e39d30c05fb648376dd713ddcbc1