Fix make test perf regression by forcing ncores = 1 (semgrep/semgrep-proprietary#3729)
The regression was introduced in
https://github.com/semgrep/semgrep-proprietary/pull/3716
because of the change to the default in Core_scan_config.ncores.
We're forcing it to 1 for osemgrep test to avoid forking.

Also removed the need for some calls to Fpath.normalize
now that the Deep diff scan code has been removed which was
doing weird things with paths.

test plan:
time make test
=> less than 5min
(was more than 10min before this PR)

synced from Pro 01fd94d7e1d6d14541fe90cbe43224ad115cb03b