feat: Upgrade support for Julia (#10820)
This PR upgrades support for Julia to a more recent version of
tree-sitter-julia.

- See tree-sitter/tree-sitter-julia#135 for the changes to the parser
- See semgrep/ocaml-tree-sitter-semgrep#520 for the changes to
semgrep-julia

Some notes:

- The changes to tree-sitter-julia in the linked PR were focused on
producing parse trees more faithful to JuliaSyntax (the default Julia
parser). However, JuliaSyntax delegates a lot of work to the lowering
phase, like checking function signatures. This means that while
tree-sitter produces a smaller parser, semgrep now needs to do more work
to find a function's name and parameters.

- Since parameters are now parsed as expressions, I left a lot of `todo`
for expressions that don't constitute valid "patterns". I don't know if
it might be a good idea to create an alias for `todo` to clearly
indicate that we don't handle those cases because they're not valid in
normal Julia (outside macros), and not because they should be handled in
the future.

- tree-sitter-julia now parses short functions as assignments (That's
also how it's done in the reference parser). To handle this, I translate
short functions into lambdas: `f(...) = ...` becomes `f = (...) -> ...`.

- The special handling for `where_clause` was removed. A later version
of tree-sitter-julia unified `where_clause` with `where_expression`.

- One of the issues I really hoped would be fixed was #10649, but that's
actually a limitation on semgrep's side! In general, I still feel
there's a lot of weird hacks because semgrep's AST only considers C-like
macros.

---------

Co-authored-by: Nat Mote <nat@natmote.net>