refactor: Take oset out of ograph interface (semgrep/semgrep-proprietary#3842)
I want to remove the oset abstraction entirely. The first step is to change the external interface and update callsites. `oset` is already represented by a `Set_` under the hood. The only change here is that we expose the underlying `Set_` to callers rather than proxying operations through the `oset` object.

We'd like to remove ograph entirely but we still have a couple of callsites so migrating off of it is likely a bit of a bigger effort.

Test plan:
* CI

synced from Pro 4f0e0ca61a168cc959956013df2ff091554e5cc3